safari-534.53-branch/LayoutTests/sputnik/Conformance/08_Types/8.4_The_String_Type/S8.4_A14_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S8.4_A14_T3

PASS Expected parsing failure

TEST COMPLETE
